media-source/SourceBuffer-remove_invalid-updating.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Calling SourceBuffer#remove(start, end) during updating is true</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = 'test.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';

function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(new Uint8Array(xhr.response));
    };
}
async_test(function(t) {
    GET(FILE, function(data) {
        var mediaSource = new MediaSource();
        mediaSource.addEventListener('sourceopen', function(e){
            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            sourceBuffer.addEventListener('updatestart', function(e) {
                t.step(function() {
                    assert_throws({name: 'InvalidStateError'},
                                  function() {
                                      sourceBuffer.remove(0, 1);
                                  },
                                  'During updating is true, calling SourceBuffer#remove(start, end)');
                });
                t.done();
            });
            sourceBuffer.appendBuffer(data);
        });
        video.src = window.URL.createObjectURL(mediaSource);
    });
},
'SourceBuffer#remove(start, end) :  During updating is true, SourceBuffer#remove(start, end) should throw InvalidStateError');
</script>
</body>
</html>
